load model in bf16 for debug

+++ nemo_rl/models/policy/dtensor_policy_worker.py
@@ -151,7 +151,7 @@ class DTensorPolicyWorker:
             # Always load the model in float32 to keep master weights in float32.
             # Keeping the master weights in lower precision has shown to cause issues with convergence.
             # https://github.com/NVIDIA/NeMo-RL/issues/279 will fix the issue of CPU OOM for larger models.
-            torch_dtype=torch.float32,
+            torch_dtype=torch.bfloat16,
             trust_remote_code=True,
             **sliding_window_overwrite(
                 model_name
